Move styles near component
