+++ app/src/components/v-emoji-picker/picmo-overrides.scss
@@ -24,8 +24,8 @@
 			--search-height: 40px;
 			--search-placeholder-color: var(--foreground-subdued);
 			--search-icon-color: var(--foreground-subdued);
-			--search-background-color: var(--background-page);
-			--search-focus-background-color: var(--background-page);
+			--search-background-color: var(--background-input);
+			--search-focus-background-color: var(--background-input);
 
 			box-shadow: var(--card-shadow);
 			border: var(--border-width) solid var(--border-normal);
