+++ app/src/styles/main.scss
@@ -13,7 +13,6 @@
 @import 'lib/highlight';
 @import 'lib/fullcalendar';
 @import 'lib/frappe-charts';
-@import 'lib/emoji-picker';
 @import 'lib/flatpickr';
 
 body.light {
